% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tbl_georeference_latlon.R
\name{tbl_georeference_latlon}
\alias{tbl_georeference_latlon}
\title{Map named latitude and longitude columns to a geographic point object.}
\usage{
tbl_georeference_latlon(DATA, cols = c("lon", "lat"), loc = "loc")
}
\arguments{
\item{DATA}{database backed dplyr tbl.}

\item{cols}{names of columns containing longitude and latitude as decimal degrees.}

\item{loc}{name of new column to be created with geographic point object.}
}
\value{
dplyr tbl object
}
\description{
Latitude and longitude coordinates must be mapped to a PostGIS object before
they can be manipulated by PostGIS functions or displayed in GIS tools. This function alters
the table referenced by \code{DATA} to add a new geography column named \code{loc} and
populates the column with the value of \code{ST_Point(lon, lat)}.
}
\note{
This function alters the database schema permanantly. If the schema has already been
modified to include the \code{loc} column then the table is returned unmodifed with a
warning.
}
\seealso{
\href{http://postgis.net/docs/ST_Point.html}{ST_Point()}
}

