% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tbl_georeference_tile.R
\name{tbl_georeference_tile}
\alias{tbl_georeference_tile}
\title{Map named slippy tile indices to a geographic point and polygon representation.}
\usage{
tbl_georeference_tile(DATA, cols = c("xindex", "yindex", "level"),
  loc = "tileloc", boundary = "tilebounds")
}
\arguments{
\item{DATA}{database backed dplyr tbl.}

\item{cols}{names of columns containing x, y, and level tile indices.}

\item{loc}{name of new column to be created with geographic point object representing the upper
left hand corner of the tile.}

\item{boundary}{name of new column to be created with the geographic polygon object representing
the tile bounday.}
}
\value{
dplyr tbl object
}
\description{
Slippy tile indices can not be displayed in GIS tools without mapping their
indices (xindex, yindex, level) to a geographic representation. This function alters the
table referenced by \code{DATA} to add two new geography columns describing the upper left
corner of the tile and the boundary of the tile.
}
\note{
This function alters the database schema permanantly. If the schema has already been
modified to include the \code{loc} or \code{boundary} column then the table is returned
unmodifed with a warning. This function also requires custom PostGIS functions from the
\href{https://github.com/petewarden/postgis2gmap/blob/master/README.md}{postgis2gmap}
project. These functions must be loaded into PostgreSQL in advance.
}

