% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tbl_tilereference.R
\name{tbl_tilereference}
\alias{tbl_tilereference}
\title{Map geographic point to slippy tile indices.}
\usage{
tbl_tilereference(DATA, loc = "loc", cols = c("xindex", "yindex", "level"),
  level = 18)
}
\arguments{
\item{DATA}{database backed dplyr tbl.}

\item{loc}{name of column containing a geographic point object to be mapped to a tile.}

\item{cols}{names of columns to be created to hold x, y, and level tile indices.}

\item{level}{numeric tile level to map to.}
}
\value{
dplyr tbl object
}
\description{
This function converts a geographic point object (\code{ST_Point(lon, lat)}) to
Slippy tile indices (xindex, yindex, level). This function alters the
table referenced by \code{DATA} to add new tile indices columns.
}
\note{
This function alters the database schema permanantly. If the schema has already been
modified to include the columns referenced by \code{cols} then the table is returned
unmodifed with a warning. This function also requires custom PostGIS functions from the
\href{https://github.com/petewarden/postgis2gmap/blob/master/README.md}{postgis2gmap}
project. These functions must be loaded into PostgreSQL in advance.
}

